refactor git
